Update the back implemetation for transactions

+++ Expenses/src/main/kotlin/br/com/money/expenses/exceptions/TransactionNotFoundException.kt
@@ -1,4 +1,4 @@
 package br.com.money.expenses.exceptions
 
-class TransactionNotFoundException(transactionId: Int) :
+class TransactionNotFoundException(transactionId: Long) :
     RuntimeException("Transaction not found with id: $transactionId")
